Update data_code.R
joined district and unique.all (Grade_7 and grade_K) into match. Still need to try and find a way to see if there are any school names that appear in only one of the sets and not the other but share a middle_6 code. Then we need to start doing ggplot work with the data and create a SBM using an R package.